Move back-button toast into scenes
